Fix apache2 not playing well with supervisor
It turns out that when you stopped the apache2 service from supervisord,
the processes were still running regardless. This fix is based on
Docker's guide to supervisord. I may change the web server in the
future, as this fix seems more like a hack to me. Maybe use a
preexisting image that can be slightly modified to accommodate
supervisord, or some other tool so that I can control the feed updater
daemon?